+++ src/com/company/TelaCriaOrcamentoTecnico.java
@@ -26,8 +26,8 @@ public class TelaCriaOrcamentoTecnico extends TelaOrdemClienteOrcamento {
         if(e.getSource() == bAprovar)
         {
             this.o.setHora(Integer.parseInt(tHora.getText()));
-            this.o.setMateriais(tMateriais.getText(), Integer.parseInt(tValMateriais.getText()));
-            this.o.setValorHora(Integer.parseInt(tValHora.getText()));
+            this.o.setMateriais(tMateriais.getText(), Float.parseFloat(tValMateriais.getText()));
+            this.o.setValorHora(Float.parseFloat(tValHora.getText()));
             this.o.settID(this.tecnico);
             this.o.setStatus("Aguardando Aprovação do Cliente");
             this.dispose();
